Move text chunking out of tts_model
